Artificial neural networks for Typhoon Modelling
This article reviews the applications of artificial neural networks (ANNs) in climate science, with a focus on typhoon modelling. The article also presents a critical analysis of various studies that have used ANNs for different aspects of typhoon modelling, such as storm track, intensity, precipitation, and surge prediction.  The article proposes a hybrid model framework that combines statistical methods and ANNs to exploit the strengths of both approaches and improve the transparency and validity of the model with specific reference to storm track prediction.